updates train_model and generate_data

+++ generate_data.py
@@ -4,12 +4,12 @@ import pandas as pd
 def generate_data(num_samples = 1000, num_features = 10, num_classes = 3):
 
     X = np.random.rand(num_samples, num_features)
-    y = np.random.randint(0, num_classes, size = num_samples)
+    y = np.random.randint(0, num_classes, size = num_samples).astype('float64')
 
     return X, y
 
 def save_data(X, y, filename_X = 'X.csv', filename_y = 'y.csv'):
-    df_X = pd.DataFrame(X)
+    df_X = pd.DataFrame(X).astype('float64')
     df_y = pd.DataFrame(y, columns=['target'])
 
     raw_data = pd.concat([df_X, df_y], axis=1)
